refactor move chart_data_apis from ChartRestApi ChartDataRestApi in seperate module
